+++ docs/changelog.rst
@@ -1,6 +1,18 @@
 Release History
 ===============
 
+v0.21.0
+-------
+
+**Disclaimer**
+
+All releases in the v0.x.x series are subject to breaking changes from one version to another. After the release of v1.0.0, this project will be subject to `semantic versioning <http://semver.org/>`_.
+
+**New Features and Enhancements**
+
+* Updated ``DoseItem.get_slice_data``, ``ImageSetItem.get_image_data``, and ``PlanItem.get_delivery_information`` methods to call into componentized RT Visualization service.
+* Corrected test for role with updated permission list.
+
 v0.20.0
 -------